refactor(calculator): improve performance and follow React best practices
Split page into server component with metadata and client component,
add React.memo to CalculatorButton and CalculatorDisplay, use functional
setState to reduce dependencies, hoist regex and use Set for O(1) lookups.
