+++ components/calculator/calculator-button.tsx
@@ -1,31 +1,29 @@
-"use client"
+"use client";
 
-import { Button } from "@/components/ui/button"
-import { cn } from "@/lib/utils"
+import { memo } from "react";
+import { Button } from "@/components/ui/button";
+import { cn } from "@/lib/utils";
 
 interface CalculatorButtonProps {
-  value: string
-  onClick: () => void
-  variant?: "default" | "secondary" | "outline"
-  className?: string
+  value: string;
+  onClick: () => void;
+  variant?: "default" | "secondary" | "outline";
+  className?: string;
 }
 
-export function CalculatorButton({
+export const CalculatorButton = memo(function CalculatorButton({
   value,
   onClick,
   variant = "outline",
-  className
+  className,
 }: CalculatorButtonProps) {
   return (
     <Button
       variant={variant}
       onClick={onClick}
-      className={cn(
-        "h-14 text-lg font-mono",
-        className
-      )}
+      className={cn("h-14 text-lg font-mono", className)}
     >
       {value}
     </Button>
-  )
-}
+  );
+});
